+++ src/Console/Command/Command.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Crunz\Console\Command;
 
 use Symfony\Component\Console\Command\Command as BaseCommand;