+++ src/Exception/CrunzException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Crunz\Exception;
 
 class CrunzException extends \Exception