+++ src/Exception/EmptyTimezoneException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Crunz\Exception;
 
-class NotImplementedException extends CrunzException
+class EmptyTimezoneException extends CrunzException
 {
 }