+++ src/Helpers.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
 use Crunz\Utils;
 
-if (!function_exists('word2number')) {
+if (!\function_exists('word2number')) {
     /**
      * Convert words to numbers.
      *
@@ -18,7 +19,7 @@ if (!function_exists('word2number')) {
     }
 }
 
-if (!function_exists('array_only')) {
+if (!\function_exists('array_only')) {
     /**
      * Get a subset of the items from the given array
      * From Illuminate/support helper functions.
@@ -36,7 +37,7 @@ if (!function_exists('array_only')) {
     }
 }
 
-if (!function_exists('setbase')) {
+if (!\function_exists('setbase')) {
     /**
      * Set the project's root directory.
      *
@@ -52,7 +53,7 @@ if (!function_exists('setbase')) {
     }
 }
 
-if (!function_exists('getbase')) {
+if (!\function_exists('getbase')) {
     /**
      * Return the project's root directory.
      *
@@ -66,7 +67,7 @@ if (!function_exists('getbase')) {
     }
 }
 
-if (!function_exists('generate_path')) {
+if (!\function_exists('generate_path')) {
     /**
      * Return absolute path for relative path.
      *
@@ -82,7 +83,7 @@ if (!function_exists('generate_path')) {
     }
 }
 
-if (!function_exists('getroot')) {
+if (!\function_exists('getroot')) {
     /**
      * Return absolute path for relative path.
      *