+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This option allow to run all tasks regardless of configured run time,
 part of issue [#11](https://github.com/lavary/crunz/issues/11) -
 PR [#120](https://github.com/lavary/crunz/pull/120) by [@PabloKowalczyk](https://github.com/PabloKowalczyk)
 
+### Removed
+- Removed `Crunz\Output\VerbosityAwareOutput` class - PR
+[#103](https://github.com/lavary/crunz/pull/103), [@PabloKowalczyk](https://github.com/PabloKowalczyk)
+
 ## 1.7.3 - 2018-06-15
 
 ### Fixed